split main and loadcell
